Rename renderers/conftest.py to testing_utils.py
conftest.py has special pytest semantics (auto-discovery for fixtures,
hooks, and plugins), but this file only contains plain utility functions
that are explicitly imported. Rename to testing_utils.py to clarify intent.
