+++ tinker_cookbook/renderers/kimi_k2_tool_declaration_test.py
@@ -7,7 +7,7 @@ from transformers import AutoTokenizer
 
 from tinker_cookbook.renderers import get_renderer
 from tinker_cookbook.renderers.base import Message, ToolSpec, ensure_text
-from tinker_cookbook.renderers.conftest import extract_token_ids
+from tinker_cookbook.renderers.testing_utils import extract_token_ids
 from tinker_cookbook.tokenizer_utils import get_tokenizer
