+++ tinker_cookbook/renderers/qwen3_test.py
@@ -19,9 +19,9 @@ from tinker_cookbook.renderers import (
     get_renderer,
 )
 from tinker_cookbook.renderers.base import ensure_list
-from tinker_cookbook.renderers.conftest import extract_token_ids
 from tinker_cookbook.renderers.qwen3 import Qwen3Renderer
 from tinker_cookbook.renderers.qwen3_5 import Qwen3_5Renderer
+from tinker_cookbook.renderers.testing_utils import extract_token_ids
 from tinker_cookbook.tokenizer_utils import get_tokenizer
